third_party/WebKit/LayoutTests/shadow-dom/v0/event-composed-path-closed-shadowroot2-expected.txt

Event.composedPath() should include only unclosed nodes.
The full event path should be (length=12):
div#target, #shadow-root (open), div#div3_open, #shadow-root (closed),
div#div2_closed, #shadow-root (open), div#host_open, div#host_closed,
body, html, #document, window

On #host_closed, #host_open, and #div2_closed,
div#target, #shadow-root (open), div#div3_open, #shadow-root (closed)
will be trimmed (length=8).


Dispaching a click event on #target, listening on #host_closed.
Got event.composedPath() for #host_closed:
#div2_closed, [object ShadowRoot], #host_open, #host_closed, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 8

Dispaching a click event on #target, listening on #host_open.
Got event.composedPath() for #host_open:
#div2_closed, [object ShadowRoot], #host_open, #host_closed, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 8

Dispaching a click event on #target, listening on #div2_closed.
Got event.composedPath() for #div2_closed:
#div2_closed, [object ShadowRoot], #host_open, #host_closed, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 8

Dispaching a click event on #target, listening on #div3_open.
Got event.composedPath() for #div3_open:
#target, [object ShadowRoot], #div3_open, [object ShadowRoot], #div2_closed, [object ShadowRoot], #host_open, #host_closed, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 12

Dispaching a click event on #target, listening on #target.
Got event.composedPath() for #target:
#target, [object ShadowRoot], #div3_open, [object ShadowRoot], #div2_closed, [object ShadowRoot], #host_open, #host_closed, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 12
PASS successfullyParsed is true

TEST COMPLETE
